omap_hsmmc: Handle sdio interrupts when clocks are disabled *PRELIMINARY*
When the mmc host disables its internal clocks, it can no longer detect
sdio interrupts on dat1. To continue being able to serve such interrupts,
dat1 should be remuxed to the relevant gpio and handled as an irq source.

The interrupt handler should be run in process context, as enable_sdio_irq
might sleep. Thus, a threaded irq is used for this purpose.

Signed-off-by: Ido Yariv <ido@wizery.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>