mmc: Toggle power while host is claimed
Runtime PM ops might be called when the host isn't claimed. This can be
problematic, as the mmc disable/enable callbacks might be called
concurrently.

Fix this by claiming the host before calling mmc_power_*_host. Since
it's perfectly fine to claim a host more than once by the same task, it
should be safe to do even if the runtime PM functions are called with
the host claimed.

Signed-off-by: Ido Yariv <ido@wizery.com>
Signed-off-by: Ido Reis <idor@ti.com>
Signed-off-by: Eliad Peller <eliad@wizery.com>