mmc: omap_hsmmc: Avoid spurious interrupts in dat1 handler
*** HACK ***

For some reason, probably because we request/free the interrupt handler each
time, we sometimes get spurious interrupts.
Some SDIO devices may enter low power mode. In this case, it might not be
possible to read the pending interrupts register, as the SDIO layer might not
be fully operational.
Check the gpio line before actually handling the interrupt.

TODO:
- Find out the root cause for the spruious interrupts
- Consider enabling/disabling the interrupt handler instead of
  requesting/freeing. This should be done regardless of this specific problem.

Signed-off-by: Ido Yariv <ido@wizery.com>