Enable Asynchronous Int featute for SDIO host
Enble the use of SDIO card interrupts.

FCLK must be enabled while SDIO interrupts are enabled or the MMC
module won't wake-up (even though ENAWAKEUP in SYSCONFIG and IWE in
HTCL have been set).  Enabling the MMC module to wake-up would require
configuring the MMC module (and the mmci_dat[1] GPIO when the CORE
power domain is OFF) as wake-up sources in the PRCM.

Conditional check is also used to set registers for
inabnd interrupt differently.

Signed-off-by: Sourav Poddar <sourav.poddar@ti.com>
[Eliad: Fix omap_hsmmc_enable_sdio_irq prototype]
Signed-off-by: Eliad Peller <eliad@wizery.com>
Signed-off-by: Ido Reis <idor@ti.com>